Moved all HTTP-related exceptions into the http package inside of azure-core.
