+++ sdk/core/azure-core/src/main/java/com/azure/android/core/http/exception/ClientAuthenticationException.java
@@ -1,7 +1,7 @@
 // Copyright (c) Microsoft Corporation. All rights reserved.
 // Licensed under the MIT License.
 
-package com.azure.android.core.exception;
+package com.azure.android.core.http.exception;
 
 import okhttp3.Response;
